Split FlagsChangeRebuildTests
